[services] Tunnelbroker - Refactor handler names in Amqp `connectInternal()`
Summary:
This diff is a part of the stack.

Introducing refactoring to the handler variable names in `AmqpManager` -> `connectInternal()` method to reflect its purposes instead of using the shorthand.

Related linear task: [[ https://linear.app/comm/issue/ENG-1495/amqp-cpp-client-throwing-a-segmentation-error-in-a-highmid-load-and | ENG-1495 ]]

Test Plan:
Successfully built using `yarn run-tunnelbroker-service-in-sandbox` command.
Passing all AMQP unit tests in the last diff D4749 in a stack.

Reviewers: karol, tomek

Reviewed By: karol, tomek

Subscribers: ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4743