[services] Clean up `install_cryptopp.sh` using ShellCheck
Summary: Related Linear issue [[ https://linear.app/comm/issue/ENG-1295/clean-up-shell-scripts-in-codebase | here ]]. This is part of a set of diffs that will allow ShellCheck to be added to the CI. See inline comments for specific details of the ShellCheck error/warning output.

Test Plan:
[[ https://www.shellcheck.net/ | ShellCheck ]]

This script is run in the `tunnelbroker` Docker build, so if that CI continues to build with no errors we can reasonably assume this script continues to work as expected.

Also, I ran `git grep 'CXXFLAGS'` and found no occurrences of it in the codebase.

Reviewers: atul, jon, tomek, karol

Reviewed By: atul, jon

Subscribers: ashoat, adrian, karol

Differential Revision: https://phab.comm.dev/D4738