[services] Tunnelbroker - Change RabbitMQ queue name for a sandbox mode
Summary:
When running in a sandbox mode Tunnelbroker uses a different config file which is `tunnelbroker-dev.ini`. In a sandbox mode, we are using a `test` prefix for database table names.
To follow this approach we should use a test name for a RabbitMQ queue instead of the default one.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1491/deploy-and-ci-testing-related-fixes | ENG-1491 ]]

Test Plan: Successfully passing unit tests from D4749.

Reviewers: karol, tomek

Reviewed By: karol, tomek

Subscribers: ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4753