[Nix] Build amqp-cpp with tcp extension on mac
Summary:
Enable tcp extensions on darwin.
This is necessary as tunnelbroker makes use of them.
Upstream considers these "able to be used by darwin,
but untested".

Since we more or less want to use this for development,
there's not much harm to having this available for mac as
it's not a production scenario.

Test Plan:
- install nix
```
# From mac machine
nix build .#amqp-cpp
```

Reviewers: varun, atul, max, karol

Subscribers: ashoat, tomek, adrian, abosh

Differential Revision: https://phab.comm.dev/D4764