Implement methods to store and retrieve current_user_id from database in CommCoreModule
Summary: This differential via CommCoreModule to JavaScript methods to retrieve and set current user ID that are defined and implemented in DatabaseQueryExecutor and SQLiteQueryExecutor in the parent diff.

Test Plan: Test plan is exactly the same as for parent diff with the difference that we want to call correspinding methods of `global.CommCoreModule` object instead of SQLiteQueryExecutor directly. This way we can be sure they work in JavaScript as well.

Reviewers: tomek

Subscribers: