[native] Fix message tooltip positioning for robotext with sidebar
Summary:
The diff fixes: https://linear.app/comm/issue/ENG-1591/messagetooltip-renders-in-wrong-place-if-robotext-message-contains
It moves inlineSidebar for robotext messages one level higher in component tree, so it won't be included when measuring message position and size - like it is with `composedMessage`.
Test Plan:
Test if toolptip displays correctly on android and iOS, confirm that animations for opening or creating sidebar modal still works as before.
Reviewers:
tomek, atul
Subscribers: