[keyserver] Create messages for open GENESIS subchannels
Summary:
To avoid spammming everybody, today we don't create messages when a subchannel of `GENESIS` is created.

But for the seed round party, we want to create an open subchannel of `GENESIS` that everybody can access. To make it easy to access, we want to have a message appear in `GENESIS` about the channel.

This diff makes it so the messages are still created for open subchannels of `GENESIS`. The only way for an open subchannel of `GENESIS` to be created is for me (@ashoat) to do it manually, so this is safe.

Test Plan: I tested this in my local environment and it created a message (but only when combined with the hacks listed [here](https://linear.app/comm/issue/ENG-1544#comment-c221fbe3))

Reviewers: atul

Reviewed By: atul

Subscribers: tomek, adrian, abosh

Differential Revision: https://phab.comm.dev/D4808