[keyserver] Introduce make-channel-private.js script
Summary:
For the seed round party, we're going to create an open subchannel of `GENESIS` that everyone on the app can see.

After the party, we want to hide this channel (make it private). This diff introduces a script that can do that. See step 6 of [this plan](https://linear.app/comm/issue/ENG-1544#comment-c221fbe3) for more details.

Depends on D4808

Test Plan: I tested this in my local environment and it immediately hid the channel for the `test` user, both in message history and in the inbox, without crashing the app or causing any issues

Reviewers: atul

Reviewed By: atul

Subscribers: tomek, adrian, abosh

Differential Revision: https://phab.comm.dev/D4809