[web] Add redux persist dependency
Summary:
We would like to use persistence on web and the most convenient way to do that is to use `redux-persist`. We already use that library on native. The default persistence location of web is local storage which is what we need.

This diff also contains a patch that fixes a bug in `redux-persist` types: `_persist` should be an optional property. I've decided to put these two changes in one diff because without the patch this library is basically unusable on web.

Test Plan: Run the web and check if opening the app works. The behavior shouldn't be affected by this change.
Check if running `yarn install` installs the library with the patch applied.

Reviewers: jacek, atul, ashoat