[native] add blocking version of get_user_id method
Summary:
Essentially converts the async get_user_id method to a sync method by blocking until the future is resolved. We do this because async functions are not allowed in the CXX bridge yet.

Depends on D4802

Test Plan: Successfully called the sync version and get a response from the Identity service.

Reviewers: tomek!, karol

Subscribers: ashoat, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4815