[iOS] Bump `NotificationService` `CFBundleVersion` to match parent app
Summary:
Was noticing this warning at the end of iOS builds:

```
warning: The CFBundleVersion of an app extension ('1') must match that of its containing parent app ('141').
```

This change gets rid of the warning.

Does it make a difference? Probably not.

Test Plan:
Did command-line build and made sure that warning didn't appear at the end of the logs.

Reviewers:
marcin, abosh, tomek, jacek, varun