[native] Replace `cloud-filled` icon in `CommunityPill` with `CommIcon`
Summary:
Searched through codebase for usages of the `cloud-filled` icon and found one instance in the `CommunityPill` component.

Swapped out the `SWMansionIcon` component for the `CommIcon` component.

> Why did the size change?

I believe it's because there was additional space/padding around the previous `cloud-filled` icon that made it appear smaller than it actually was. Will double check this once I get to re-creating the `SWMansionIcon` pack with **all** of the provided icons and **Without** the custom Comm ones.

Note: Going to create individual diffs for all usages of each icon for `CommIcon`s to make the before/after clear. There are only 6 custom `CommIcon`s so this shouldn't produce too much spam.

On the other hand going to do all of the `SWMansion` icons together in one go.

Test Plan:
Before:
After:

Reviewers:
jacek, tomek, abosh, varun