[native] Remove `sortBy` property from `fetchPhotos`
Summary: Relevant Linear issue [[ https://linear.app/comm/issue/ENG-1481/revisit-medialibrary-sortby | here ]]. I left detailed research for this diff in my comment there. This diff sorts out issues with `MediaGalleryKeyboard` displaying images in a weird order (no pun intended) and instead mirrors the "default sorting provided by the platform" (from the [[ https://docs.expo.dev/versions/latest/sdk/media-library/#assetsoptions | Expo docs ]]).

In summary, photos are now displayed in the same order as on a user's Camera Roll.

Test Plan: Tested on physical iOS device and Android simulator and works as expected. One notable thing I tested was setting the `creationTime` of a photo to the future and making sure that it wasn't pushed to the front of the photos displayed on `MediaGalleryKeyboard`. This was the main bug with using `creationTime`.

Reviewers: atul, tomek

Subscribers: