[native] Replace `sidebar-filled` icon in `SwipeableMessage` with `CommIcon`
Summary:
Searched through codebase for usages of the `sidebar-filled` icon and found one instance in the `SwipeableMessage` component.

Swapped out the `SWMansionIcon` component for the `CommIcon` component.

NOTE: Going to create individual diffs for all usages of each icon for CommIcons to make the before/after clear. There are only 6 custom CommIcons so this shouldn't produce too much spam.

On the other hand going to do all of the SWMansion icons together in one go.

Test Plan:
Before:
{F140044}

After:
{F140045}

Reviewers: jacek, tomek, abosh, varun

Reviewed By: abosh, varun

Subscribers: ashoat, adrian

Differential Revision: https://phab.comm.dev/D4840