[web] Replace `cloud-filled` icon in `ChatThreadAncestors` with `CommIcon`
Summary:
Searched through codebase for usages of `cloud-filled` and found one instance in the `ChatThreadAncestors` component.

Swapped out the `SWMansionIcon` component for the `CommIcon` component.

NOTE: Going to create individual diffs for all usages of each icon for CommIcons to make the before/after clear. There are only 6 custom CommIcons so this shouldn't produce too much spam.

Test Plan:
WARNING: The actual icon changed. I'm using the same cloud icon we decided with Alicja to go with on `native`. I strongly think we should go with the same filled cloud icon on web because A. consistency B. it looks a lot better than an open path.

Before:

{F140471}

After:

{F140472}

Reviewers: jacek, tomek, abosh, varun

Reviewed By: abosh, varun

Subscribers: ashoat, adrian

Differential Revision: https://phab.comm.dev/D4852