[native] helper to handle access token response
Summary:
The last message sent by the Identity service in the registration workflow contains an access token. We make sure that we have the right message type and return the token in this handler.

Depends on D4856

Test Plan: `cargo build`. this will be tested by the method that uses the helper later in the stack

Reviewers: tomek!, karol, atul

Subscribers: