[web] Update `SWMansionIcon` component with cleaned up icons
Summary:
Same as D4865, but for `web` instead of `native`.

Copied from D4865:
> It took a **while** to clean up the source vectors (converting strokes to outlines, joining paths w/ union, subtract, intersect, difference properly, flattening things, getting alignment correct, getting things exported properly, etc).
>
> Once that was all figured out, imported the cleaned up icons into IcoMoon and prepared for export. Retrieved the relevant TTF fonts and JSON config from IcoMoon and replaced the existing fonts/config in the project.
>
> Everything continues to look as expected (or better), but a few icon names have changed. I'll add the icon name changes to this diff to avoid spamming Phabricator with a bunch of rename diffs.

We're using the same `swmansion-icon-config.json` file exported from IcoMoon on both `web` and `native` now which is great because it means icons across platforms are "backed" by the same source vectors and kept in sync.

---

We're now including the **entire** SWMansion icon pack (all 275 icons) so we should hopefully never have to look at this again. If we have additional icons we want to include in the future, they'll live in (the much smaller) `CommIcons` pack.

Test Plan:
WARNING: The icons have some visual artifacts (same as encountered in D4864) that are fixed in D4874.

Tested the `web` app in Safari/Chrome and the icons appeared as expected. There are a handful of icons that don't appear because their names have changed... I went through and replaced the old icon names with the corresponding new ones.

But other than that, it looks like this change went smoothly.

Reviewers: abosh, varun, tomek, jacek

Reviewed By: abosh, varun

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D4866