[web] Update `SWMansionIcon` component with cleaned up icons
Same as D4865, but for `web` instead of `native`.

Copied from D4865:
> It took a **while** to clean up the source vectors (converting strokes to outlines, joining paths w/ union, subtract, intersect, difference properly, flattening things, getting alignment correct, getting things exported properly, etc).
>
> Once that was all figured out, imported the cleaned up icons into IcoMoon and prepared for export. Retrieved the relevant TTF fonts and JSON config from IcoMoon and replaced the existing fonts/config in the project.
>
> Everything continues to look as expected (or better), but a few icon names have changed. I'll add the icon name changes to this diff to avoid spamming Phabricator with a bunch of rename diffs.

We're using the same `swmansion-icon-config.json` file exported from IcoMoon on both `web` and `native` now which is great because it means icons across platforms are "backed" by the same source vectors and kept in sync.

NOTE: While most icons work as they did before (or better), there are a few icons that have had their names changed. I'll do a find/replace in the morning to clean up any discrepancies and update the `flow` types.

Test Plan:
Tested the `web` app in Safari/Chrome and the icons appeared as expected. There are a handful of icons that don't appear because their names have changed... I'll go through and replace the old icon names with the corresponding new ones and update this diff.

But other than that, it looks like this change went smoothly.

Reviewers:
abosh, varun, tomek, jacek