[native] Clean up lines in `retryMultimediaMessage`
Summary: More fixes I saw while reading this for video work.

Test Plan:
For the nullish coalescing operator, it's the same reasoning as D4873: `pendingUploads` is an object, so `!pendingUploads` will only be true when `pendingUploads` is `null` or `undefined`. Thus the nullish coalescing operator covers these cases.

The removal of the `else` is because of no `else` after `return`.

Reviewers: atul

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D4877