[native] Clean up lines in `retryMultimediaMessage`
Summary: More fixes I saw while reading this for video work.

Test Plan: For the nullish coalescing operator, it's the same reasoning as D4873: `pendingUploads` is an object, so it'll `!pendingUploads` will only be true when `pendingUploads` is `null` or `undefined`. So the nullish coalescing operator covers these cases.

The removal of the `else` is because of no `else` after `return`.

Reviewers:

Subscribers: