[native] Make `failed` `boolean` in `PendingMultimediaUpload`
Summary: Change `failed` from `?string` to `boolean` in PendingMultimediaUpload`. This diff also adds an optional `failureMessage` prop to `PendingMultimediaUpload` to store the error message in case `failed` is `true`.

This is part of an effort to refactor `InputStateContainer` since some of its logic is confusing. More context [[ https://phab.comm.dev/D4877#141648 | here ]].

Test Plan: Made sure sending photos and videos worked as expected on iOS Simulat
or and added `console.log(...)` statements to log `pendingUploads` to view `failed` as a `boolean`.

Reviewers: atul, tomek, jacek

Subscribers: