[services] Tunnelbroker - Fix `amqpReady` assignment
Summary:
This diff is a part of the stack.

The private variable `amqpReady` in AmqpManager reflects the state of the AMQP Client connection state.
To update the state we must use only two callbacks:
- [[ https://github.com/CopernicaMarketingSoftware/AMQP-CPP/blob/db6f2f23dca8ebd0dcccb9d4bf6b102c33c00840/include/amqpcpp/channel.h#L81 | OnReady() ]]
- [[ https://github.com/CopernicaMarketingSoftware/AMQP-CPP/blob/db6f2f23dca8ebd0dcccb9d4bf6b102c33c00840/include/amqpcpp/channel.h#L95 | OnError() ]]

This diff introduces assignments on `OnReady()` and removes the unneeded assignments.

Related linear task: [[ https://linear.app/comm/issue/ENG-1495/amqp-cpp-client-throwing-a-segmentation-error-in-a-highmid-load-and | ENG-1495 ]]

Test Plan:
Successfully built using `yarn run-tunnelbroker-service-in-sandbox` command.
Passing all AMQP unit tests in the last diff D4749 in a stack.

Reviewers: karol, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4742