Implement utility function in WorkerThread class to block temporarily block thread execution once all scheduled tasks complete
Summary: The reason to introduce this diff is that there are react components that might try to access database while its being deleted or recreated after deletion. Such scenario has not been observed yet but it is probably because the process happens so fast so it is race condition. To avoid a crash when a component tries to access SQLite while it is being deleted we must temporarily block databaseThread argument of CommCoreModule. This differential introduces API to do so. Firstly it synchronously waits unitl all tasks in the queue complete. Then it schedules a task that block a thread execution until a promies value is set. The promise that controls how long database thread is suspended is returned to the caller.

Test Plan: The best way to test is to use this function somwhere with log statements. The order of logs should be predictable

Reviewers: tomek

Subscribers: