[keyserver] Add `uploadResult` to `uploadRows` object in `createUploads`
Summary:
Quick change. This is more concise because `uploadRows` already had a `insert` property to mark the array of values inserted into the `uploads` table, but since `uploadRows` was also used to create the `UploadMultimediaResult[]` return value of `createUploads` the rest of the object had to be remapped in the return statement of `createUploads`.

This diff adds the `uploadResult` property to `uploadRows` when the `uploadInfos` are first mapped (by copying the same code that was mapping `uploadRows` in the return statement) so that the return statement is less redundant and can just map each element of `uploadRows` to `uploadResult`.

Test Plan:
`console.log(...)`'d the output of `createUploads` before and after making this change, making sure the outputs matched for the same image and video uploads.

Also verified that the `uploads` table looked the same for an image uploaded before and after this change.

Reviewers: atul, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D4898