[native] "Flip the switch" to enable video messages
Summary:
Now that we have video messages working end-to-end, we can "flip the switch" to enable video messages.

Specifically:
1. Set `serverCanHandle` to `true` for video types in `lib/media/file-utils`
2. Set versionCode for `hasMinCodeVersion` check in `multimedia-message-spec` to 148
3. Add `MediaLibrary.MediaType.video` to `mediaType` array in `multimedia-gallery-keyboard.react` so users can select videos from their media gallery

---

Depends on D5240

Test Plan:
All the diffs that came before this one in the stack tested various part of the video message system.

Outside of that I've done a lot of testing on iOS/Android and different codeVersions to make sure that things work as expected... and I'm going to do a lot more before I land this diff.

Reviewers:
tomek, marcin, abosh, ashoat

Subscribers: