[services] Tunnelbroker - Amqp Manager in parallel threads messages throughput test
Summary:
This diff introduces AMQP Manager and AMQP-C++ client implementation multithreaded access test. In this test, we are spawning 100 threads that send 10 messages each in parallel. In this case, we are checking for possible deadlocks and unsynchronized access to the shared resources inside the AMQPManager implementation.

Also, this test can show the throughput performance of the single shared AMQP channel implementation to send/receive messages.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1495/amqp-cpp-client-throwing-a-segmentation-error-in-a-highmid-load-and | ENG-1495 ]]

Test Plan: Run `yarn run-unit-tests tunnelbroker` command and test is successfully passed.

Reviewers: karol, tomek

Reviewed By: karol, tomek

Subscribers: ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4768