[services] Add curl to the base image
Summary:
Depends on D4858

I added `curl` to the base image since it is needed to install rust.

The reason I didn't add rust right away to the base image is that we'll be using rust in the backup and tunnelbroker which makes it redundant in the blob service.

Test Plan: services build

Reviewers: tomek, jon

Reviewed By: tomek, jon

Subscribers: ashoat, tomek, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4859