[services] Tunnelbroker - Add copying of the shared directory to the Dockerfile
Summary:
This diff introduces changes to the `Dockerfile` to copy the `/shared` directory to the service Docker container.
In a [[ https://linear.app/comm/issue/ENG-1698/moving-of-protos-cpp-and-rust-shared-code-into-a-root-shared-folder | ENG-1698 ]] we have introduced a new directory for the shared objects.

To make the Tunnelbroker built in D4807 and beyond updates, we need to copy the `/shared` folder to the Docker container.

Related Linear tasks: [[ https://linear.app/comm/issue/ENG-1698/moving-of-protos-cpp-and-rust-shared-code-into-a-root-shared-folder | ENG-1698 ]], [[ https://linear.app/comm/issue/ENG-1672/corrosion-cxx-integration-for-services | ENG-1672 ]]

Test Plan: Successfully built in a D4807 using the `cmake/corrosion-cxx.cmake ` file from the `shared` directory.

Reviewers: karol, tomek, jon

Reviewed By: karol, tomek

Subscribers: ashoat, tomek, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4949